requirements.txt
Django>=2.2
djangorestframework
six>=1.4.1
unicodecsv
